refactor: move user update listen to new purgehooks api
