+++ utils/caching/caching_strategy.md
@@ -8,8 +8,8 @@ This document briefly describes the caching strategy employed on PubPub.
 
 PubPub caches pages on Fastly that follow the following rules:
 
--   Do not match `req.url ~ "(/login|/logout|/api|/auth|\?access=)"`
--   Do not match `req.http.Cookie ~ "pp-no-cache"`. The `pp-no-cache` cookie is set when you log in.
+-   Do not match `req.url ~ "(/login|/logout|/auth|\?access=)" || req.url ~ "^/api/(pubs/.+/text|logout|userNotifications|uploadPolicy|citations|users|search/users|zoteroIntegration|editor|generateDoi|pubEdgeProposal|pubEdges|pubHistory|workerTasks)"`
+-   Do not match `req.http.Cookie ~ "pp-lo"`. The `pp-lic = pp-lo` cookie is set when your'e logged out.
 
 This means that public facing pages for non-logged in users are cached, as those make up the majority of traffic.
 
@@ -46,7 +46,11 @@ While not very precise, this makes sure we do invalidate the cache when necessar
 
 Purging happens in four places
 
--   The purge middleware defined here: https://github.com/pubpub/pubpub/blob/master/utils/caching/purgeMiddleware.ts#L1-L100
+#### Purge middleware
+
+We have a middleware that is called on every request that checks if the request is a POST/PUT/PATCH/DELETE request and if so, purges the cache for the community the request is for.
+
+It is defined here: https://github.com/pubpub/pubpub/blob/master/utils/caching/purgeMiddleware.ts#L1-L100
 
 #### After workerTask completion
 
@@ -58,9 +62,17 @@ If we purged only after the `/api/release` or `/api/export` task had been called
 
 https://github.com/pubpub/pubpub/blob/master/workers/queue.ts#L84-L86
 
-#### After a PubAttribution is created
+#### Certain Model lifecycle hooks
 
-This is to counter a kind of niche situation.
+Sometimes you might want to purge the cache when a certain model is created, updated or destroyed.
+
+To do this you can use a very similar API as the activityHooks, called `createPurgeHooks`.
+
+Almost all the current uses of this are to update the `/user/*` pages, or other places where the user is shown.
+
+See https://github.com/pubpub/pubpub/blob/master/utils/caching/createPurgeHooks.ts
+
+##### After PubAttribution is created, or a Release is created
 
 Situation:
 
@@ -79,16 +91,34 @@ This does not need to happen when a PubAttribution is updated or destroyed, the 
 -   A PubAttribution is updated or destroyed on `demo.pubpub.org`, sending a purge request for that
 -   Barbara's user page is now uncached
 
-#### After PUT /api/users is called
+###### Even more specific case
 
-This is technically a subset of the middleware, but is kind of special.
+The above is valid when a user is added to an already published Pub, but we also need to purge when a Pub is published for the first time.
+
+Problem workflow:
+
+1. Add user which has no pubs in this community to a pub. This sends a purge request to the user page
+2. Go to www.pubpub.org/user/<user-slug>. Cache miss, but no pub, because it isn't released.
+3. Release Pub
+4. Go back to www.pubpub.org/user/<user-slug>. Still no Pub, because this page isn't purged on release.
+5. Remove and re-add user
+6. Go back to www.pubpub.org/user/<user-slug>. Now the pub shows up.
+
+##### User is modified
 
 When a user updates their profile, we expect this to reflected anywhere their profile is visible.
 E.g., imagine Barbara from before first did not have a profile picture, and now adds one.
 
-### Purge middleware
+This needs to be updated in the following places:
 
-We have a middleware that is called on every request that checks if the request is a POST/PUT/PATCH/DELETE request and if so, purges the cache for the community the request is for.
+1. On every /user/barbara page, as the profile picture is shown there.
+2. On every page where Barbara is shown as a member.
+3. On every pub where Barbara is shown as an author.
+4. On every logged-in page for Barbara themselves, they need to see their profile pic updated.
+
+#### User notifications
+
+TODO
 
 ### Debouncing
