+++ utils/caching/purgeMiddleware.ts
@@ -1,11 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import { defer } from 'server/utils/deferred';
-import { sequelize } from 'server/models';
-import { QueryTypes } from 'sequelize';
 import type { createCachePurgeDebouncer } from './schedulePurge';
-import { uniqueCommunitiesFromMembersQuery } from './uniqueCommunitiesFromMembersQuery';
-import { getCorrectHostname } from './getCorrectHostname';
-import { getPPLic } from './getHashedUserId';
 import { shouldntPurge } from './skipPurgeConditions';
 
 const ALLOWED_METHODS = ['POST', 'PUT', 'DELETE', 'PATCH'] as const;
@@ -38,17 +33,6 @@ const userPaths = [
 	'/api/userNotificationPreferences',
 	'/api/userDismissable',
 ] as const;
-
-async function getUniqueHostnamesForUserId(userId: string) {
-	const result = (await sequelize.query(uniqueCommunitiesFromMembersQuery, {
-		replacements: { userId },
-		type: QueryTypes.SELECT,
-	})) as { domain: string | null; subdomain: string }[];
-
-	const hostnames = result.map(({ domain, subdomain }) => getCorrectHostname(subdomain, domain));
-
-	return hostnames;
-}
 
 /**
  * These are routes with path parameters that we don't want to purge
@@ -126,29 +110,11 @@ export const purgeMiddleware = (
 			}
 
 			defer(async () => {
-				if (req.path === '/api/users' && req.method === 'PUT') {
-					const hostnames = await getUniqueHostnamesForUserId(req.user.id);
-
-					const allPurges = [
-						// all the communities the user is a member of,
-						// either directly or through a pub/collection
-						...hostnames,
-						// the current community
-						surrogateTag,
-						// purge the cache for the logged in user
-						getPPLic(req.user),
-						// purge all the /user pages
-						req.user.slug,
-					];
-
-					await Promise.all(allPurges.map(async (tag) => schedulePurge(tag)));
-				} else {
-					/**
-					 * this await is here so you can wait on the purge requests
-					 * using `finishDeferredTasks` in tests or other short lived processes
-					 */
-					await schedulePurge(surrogateTag);
-				}
+				/**
+				 * this await is here so you can wait on the purge requests
+				 * using `finishDeferredTasks` in tests or other short lived processes
+				 */
+				await schedulePurge(surrogateTag);
 			});
 		});
